requirements.txt
fastapi
libsass
pandas
uvicorn
python-multipart
jinja2
pandas
seaborn
psycopg2-binary
fuzzywuzzy
scikit-learn
requests
python-dotenv
sqlalchemy